[Routing] Refactoring
Move basic functions from turns_generator to turns_generator_utils.
