+++ routing/turns_generator.cpp
@@ -1,4 +1,5 @@
 #include "routing/turns_generator.hpp"
+#include "routing/turns_generator_utils.hpp"
 
 #include "routing/router.hpp"
 #include "platform/measurement_utils.hpp"
@@ -44,18 +45,6 @@ int constexpr kMinHighwayClassDiffForService = -1;
 // Min difference between HighwayClasses of the route segment and alternative turn segment
 // to ignore this alternative candidate (when route direction is GoStraight).
 int constexpr kMinHighwayClassDiffForStraightRoute = -1;
-
-bool IsHighway(HighwayClass hwClass, bool isLink)
-{
-  return (hwClass == HighwayClass::Trunk || hwClass == HighwayClass::Primary) &&
-         !isLink;
-}
-
-bool IsSmallRoad(HighwayClass hwClass)
-{
-  return hwClass == HighwayClass::LivingStreet ||
-         hwClass == HighwayClass::Service || hwClass == HighwayClass::Pedestrian;
-}
 
 /// \brief Return |CarDirection::ExitHighwayToRight| or |CarDirection::ExitHighwayToLeft|
 /// or return |CarDirection::None| if no exit is detected.
@@ -110,13 +99,6 @@ CarDirection TryToGetExitDirection(TurnCandidates const & possibleTurns, TurnInf
                                             : CarDirection::ExitHighwayToLeft;
   }
   return CarDirection::None;
-}
-
-/// * \returns difference between highway classes.
-/// * It should be considered that bigger roads have smaller road class.
-int CalcDiffRoadClasses(HighwayClass const left, HighwayClass const right)
-{
-  return static_cast<int>(left) - static_cast<int>(right);
 }
 
 /// * \returns true when
@@ -282,32 +264,6 @@ bool FixupLaneSet(CarDirection turn, vector<SingleLaneInfo> & lanes,
     }
   }
   return isLaneConformed;
-}
-
-/*!
- * \brief Converts a turn angle into a turn direction.
- * \note lowerBounds is a table of pairs: an angle and a direction.
- * lowerBounds shall be sorted by the first parameter (angle) from big angles to small angles.
- * These angles should be measured in degrees and should belong to the range [-180; 180].
- * The second paramer (angle) shall belong to the range [-180; 180] and is measured in degrees.
- */
-template <class T>
-T FindDirectionByAngle(vector<pair<double, T>> const & lowerBounds, double angle)
-{
-  ASSERT_GREATER_OR_EQUAL(angle, -180., (angle));
-  ASSERT_LESS_OR_EQUAL(angle, 180., (angle));
-  ASSERT(!lowerBounds.empty(), ());
-  ASSERT(is_sorted(lowerBounds.rbegin(), lowerBounds.rend(), base::LessBy(&pair<double, T>::first)),
-         ());
-
-  for (auto const & lower : lowerBounds)
-  {
-    if (angle >= lower.first)
-      return lower.second;
-  }
-
-  ASSERT(false, ("The angle is not covered by the table. angle = ", angle));
-  return T::None;
 }
 
 RoutePointIndex GetFirstOutgoingPointIndex(size_t outgoingSegmentIndex)
@@ -562,17 +518,6 @@ double CalcRouteDistanceM(vector<geometry::PointWithAltitude> const & junctions,
 bool RoutePointIndex::operator==(RoutePointIndex const & index) const
 {
   return m_segmentIndex == index.m_segmentIndex && m_pathIndex == index.m_pathIndex;
-}
-
-// TurnInfo ----------------------------------------------------------------------------------------
-bool TurnInfo::IsSegmentsValid() const
-{
-  if (m_ingoing->m_path.empty() || m_outgoing->m_path.empty())
-  {
-    LOG(LWARNING, ("Some turns can't load the geometry."));
-    return false;
-  }
-  return true;
 }
 
 bool GetNextRoutePointIndex(IRoutingResult const & result, RoutePointIndex const & index,
@@ -896,70 +841,6 @@ CarDirection GetRoundaboutDirection(TurnInfo const & turnInfo, TurnCandidates co
   bool const keepTurnByHighwayClass = KeepRoundaboutTurnByHighwayClass(nodes, turnInfo, numMwmIds);
   return GetRoundaboutDirection(turnInfo.m_ingoing->m_onRoundabout, turnInfo.m_outgoing->m_onRoundabout, 
                                 nodes.candidates.size() > 1, keepTurnByHighwayClass);
-}
-
-CarDirection InvertDirection(CarDirection dir)
-{
-  switch (dir)
-  {
-    case CarDirection::TurnSlightRight:
-      return CarDirection::TurnSlightLeft;
-    case CarDirection::TurnRight:
-      return CarDirection::TurnLeft;
-    case CarDirection::TurnSharpRight:
-      return CarDirection::TurnSharpLeft;
-    case CarDirection::TurnSlightLeft:
-      return CarDirection::TurnSlightRight;
-    case CarDirection::TurnLeft:
-      return CarDirection::TurnRight;
-    case CarDirection::TurnSharpLeft:
-      return CarDirection::TurnSharpRight;
-    default:
-      return dir;
-  };
-}
-
-CarDirection RightmostDirection(const double angle)
-{
-  static vector<pair<double, CarDirection>> const kLowerBounds = {
-      {145., CarDirection::TurnSharpRight},
-      {50., CarDirection::TurnRight},
-      {10., CarDirection::TurnSlightRight},
-      // For sure it's incorrect to give directions TurnLeft or TurnSlighLeft if we need the rightmost turn.
-      // So GoStraight direction is given even for needed sharp left turn when other turns on the left and are more sharp.
-      // The reason: the rightmost turn is the most straight one here.
-      {-180., CarDirection::GoStraight}};
-
-  return FindDirectionByAngle(kLowerBounds, angle);
-}
-
-CarDirection LeftmostDirection(const double angle)
-{
-  return InvertDirection(RightmostDirection(-angle));
-}
-
-CarDirection IntermediateDirection(const double angle)
-{
-  static vector<pair<double, CarDirection>> const kLowerBounds = {
-      {145., CarDirection::TurnSharpRight},
-      {50., CarDirection::TurnRight},
-      {10., CarDirection::TurnSlightRight},
-      {-10., CarDirection::GoStraight},
-      {-50., CarDirection::TurnSlightLeft},
-      {-145., CarDirection::TurnLeft},
-      {-180., CarDirection::TurnSharpLeft}};
-
-  return FindDirectionByAngle(kLowerBounds, angle);
-}
-
-PedestrianDirection IntermediateDirectionPedestrian(const double angle)
-{
-  static vector<pair<double, PedestrianDirection>> const kLowerBounds = {
-      {10.0, PedestrianDirection::TurnRight},
-      {-10.0, PedestrianDirection::GoStraight},
-      {-180.0, PedestrianDirection::TurnLeft}};
-
-  return FindDirectionByAngle(kLowerBounds, angle);
 }
 
 void RemoveUTurnCandidate(TurnInfo const & turnInfo, NumMwmIds const & numMwmIds, std::vector<TurnCandidate> & turnCandidates)
